+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 ## Unreleased
 
+### Changed
+- Updated `league/oauth2-google` to 4.0.
+
 ### Fixed
 - Fixed a bug where the `site` query param was present in the OAuth redirect URI.
 - Fixed english translations.